Reestruturacao dos arquivos js e instalacao do Axios
