+++ package.json
@@ -15,5 +15,8 @@
         "laravel-mix": "^6.0.49",
         "lodash": "^4.17.19",
         "postcss": "^8.1.14"
+    },
+    "dependencies": {
+        "axios": "^1.7.7"
     }
 }
