+++ resources/js/state.js
@@ -1,0 +1,3 @@
+export const state = {
+    map: null,
+};
